Move daemon/registry to its new home at registry
