+++ daemon/daemon.go
@@ -12,11 +12,11 @@ import (
 	"github.com/manifoldco/torus-cli/base64"
 	"github.com/manifoldco/torus-cli/config"
 	"github.com/manifoldco/torus-cli/identity"
+	"github.com/manifoldco/torus-cli/registry"
 
 	"github.com/manifoldco/torus-cli/daemon/crypto"
 	"github.com/manifoldco/torus-cli/daemon/db"
 	"github.com/manifoldco/torus-cli/daemon/logic"
-	"github.com/manifoldco/torus-cli/daemon/registry"
 	"github.com/manifoldco/torus-cli/daemon/session"
 	"github.com/manifoldco/torus-cli/daemon/socket"
 )
