+++ registry/claims.go
