+++ registry/claimtree.go
