+++ registry/client.go
