+++ registry/credentialgraph.go
