+++ registry/credentials.go
