+++ registry/keypairs.go
