+++ registry/keyring.go
