+++ registry/self.go
